Removed tar remains. Better manager/server fixtures shutdown.

+++ Source/Test/test_compile.py
@@ -4,15 +4,24 @@ import sys
 import subprocess
 import tempfile
 import shutil
+import signal
 
 MGR_PORT=33221
 SRV_PORT=33222
+
+def create_file(filename):
+    return os.fdopen(os.open(filename, os.O_CREAT | os.O_WRONLY | os.O_NOINHERIT), 'w')
+
+def terminate_proc(proc):
+    proc.kill()
+    proc.communicate()
+    return True
 
 @pytest.fixture(scope='module')
 def run_manager(request):
     dir = tempfile.mkdtemp()
     ini_file = os.path.join(dir, 'distribute_manager.ini')
-    with open(ini_file, 'w') as ini:
+    with create_file(ini_file) as ini:
         ini.write("""\
 [Manager]
 port={}
@@ -23,11 +32,10 @@ node[0]=localhost:{}:4
 """.format(MGR_PORT, SRV_PORT))
     mgr_script = os.path.normpath(os.path.join(os.path.dirname(
         os.path.realpath(__file__)), '..', 'distribute_manager.py'))
-    proc = subprocess.Popen([sys.executable, mgr_script], cwd=dir, shell=True)
+    proc = subprocess.Popen([sys.executable, mgr_script], cwd=dir, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     def teardown():
-        proc.terminate()
-        proc.wait()
-        shutil.rmtree(dir)
+        if terminate_proc(proc):
+            shutil.rmtree(dir)
     request.addfinalizer(teardown)
     return proc
 
@@ -35,24 +43,24 @@ node[0]=localhost:{}:4
 def run_server(request):
     dir = tempfile.mkdtemp()
     ini_file = os.path.join(dir, 'distribute_server.ini')
-    with open(ini_file, 'w') as ini:
+    with create_file(ini_file) as ini:
         ini.write("""\
 [Server]
 port={}
 """.format(SRV_PORT))
     srv_script = os.path.normpath(os.path.join(os.path.dirname(
         os.path.realpath(__file__)), '..', 'distribute_server.py'))
-    proc = subprocess.Popen([sys.executable, srv_script], cwd=dir, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+    proc = subprocess.Popen([sys.executable, srv_script], cwd=dir, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     def teardown():
-        proc.terminate()
-        proc.wait()
-        shutil.rmtree(dir)
+        if terminate_proc(proc):
+            shutil.rmtree(dir)
     request.addfinalizer(teardown)
     return proc
 
 @pytest.fixture(scope='module')
 def db_cl():
-    return os.path.join('..', 'db_cl.exe')
+    return os.path.normpath(os.path.join(os.path.dirname(
+        os.path.realpath(__file__)), '..', 'db_cl.exe'))
 
 @pytest.fixture(scope='module')
 def vcvarsall():
@@ -83,10 +91,10 @@ def test_relative(tmpdir, run_server, run_manager, vcvarsall, db_cl):
     tmpdir = str(tmpdir)
     cpp_file = os.path.join(tmpdir, 'aaa', 'a.cpp')
     os.makedirs(os.path.dirname(cpp_file))
-    with open(cpp_file, 'w') as cpp:
+    with create_file(cpp_file) as cpp:
         cpp.write('#include "../rel.hpp"\n')
     hpp_file = os.path.join(tmpdir, 'rel.hpp')
-    with open(hpp_file, 'w') as hpp:
+    with create_file(hpp_file) as hpp:
         hpp.write('\n')
     
     env = os.environ
@@ -96,18 +104,24 @@ def test_relative(tmpdir, run_server, run_manager, vcvarsall, db_cl):
         stdout, stderr = proc.communicate()
         assert proc.returncode == 0
 
-def test_compile(tmpdir, run_server, run_manager, vcvarsall, db_cl):
-    cpp_file = os.path.join(str(tmpdir), 'a.cpp')
-    with open(cpp_file, 'w') as cpp:
+def test_link(tmpdir, run_server, run_manager, vcvarsall, db_cl):
+    tmpdir = str(tmpdir)
+    cpp_file = os.path.join(tmpdir, 'linkme.cpp')
+    with create_file(cpp_file) as cpp:
         cpp.write("int main() {}\n")
     env = os.environ
     env.update({'DB_MGR_PORT' : str(MGR_PORT)})
-    with subprocess.Popen([vcvarsall, '&&', db_cl, '/EHsc', cpp_file, "/link", "/SUBSYSTEM:CONSOLE"],
-        stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=env) as proc:
+    with subprocess.Popen([vcvarsall, '&&', db_cl, '/EHsc', cpp_file, "/link", "/SUBSYSTEM:CONSOLE", "/OUT:a_dist.exe"],
+        stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=env, cwd=tmpdir) as proc:
         stdout, stderr = proc.communicate()
         assert proc.returncode == 0
+    assert os.path.exists(os.path.join(tmpdir, 'a_dist.exe'))
 
-    
+    with subprocess.Popen([vcvarsall, '&&', 'cl', '/EHsc', cpp_file, "/link", "/SUBSYSTEM:CONSOLE", "/OUT:a_local.exe"],
+        stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=env, cwd=tmpdir) as proc:
+        stdout, stderr = proc.communicate()
+        assert proc.returncode == 0
+    assert os.path.exists(os.path.join(tmpdir, 'a_local.exe'))
 
-
+    assert os.stat(os.path.join(tmpdir, 'a_local.exe')).st_size == os.stat(os.path.join(tmpdir, 'a_dist.exe')).st_size
